Extract remembered-username helpers from LoginModal.
Keeps shell silent-restore from importing a presentational component for localStorage access.
