+++ packages/frontend/src/components/LoginModal.tsx
@@ -5,6 +5,11 @@ import {
   probeCredentials,
 } from '../api/credentialsClient'
 import { useModalKeydownFocusTrap } from '../lib/modalKeydownFocusTrap'
+import {
+  clearRememberedLoginUsername,
+  readRememberedLoginUsername,
+  writeRememberedLoginUsername,
+} from '../lib/rememberedLoginUsername'
 import { restoreFocusToElementOrFallback } from '../lib/restoreFocus'
 import { useSessionStore } from '../stores/session'
 import { cn } from '../lib/utils'
@@ -16,33 +21,6 @@ type LoginModalProps = {
   /** Called after login exchange or name-only identity switch succeeds. */
   onIdentityEstablished?: () => void
   reportShellError?: (message: string) => void
-}
-
-/** Last successful login name only (never the password). */
-export const LAST_LOGIN_USERNAME_STORAGE_KEY = 'planetsConsoleLastLoginUsername'
-
-export function readRememberedLoginUsername(): string {
-  try {
-    return localStorage.getItem(LAST_LOGIN_USERNAME_STORAGE_KEY)?.trim() ?? ''
-  } catch {
-    return ''
-  }
-}
-
-export function writeRememberedLoginUsername(username: string): void {
-  try {
-    localStorage.setItem(LAST_LOGIN_USERNAME_STORAGE_KEY, username)
-  } catch {
-    // ignore quota / private mode
-  }
-}
-
-export function clearRememberedLoginUsername(): void {
-  try {
-    localStorage.removeItem(LAST_LOGIN_USERNAME_STORAGE_KEY)
-  } catch {
-    // ignore
-  }
 }
 
 /** Match non-autofill styling; browsers force autofill backgrounds unless overridden. */
